Update README: title, overview, controls, and team members
Changed the game title to "Journey-to-Freedom: A Family’s Fight for Survival". Added an overview section in Hebrew describing the game's premise and objectives. Modified the "Demo version" section to include the game's name. Updated the controls section to remove Arrow Keys, leaving only W, A, S, D. Added a new link to a components page on GitHub. Added a team members section listing Elyasaf Kopel and Roni Levi in Hebrew.